+++ packages/components/organisms/f-footer/src/components/Footer.vue
@@ -45,7 +45,7 @@
 
         <div
             :class="['c-footer-container c-footer-row c-footer-row--combined c-footer-row--notEqualTopAndBottomPad c-footer-row--noPadBelowWide',
-                     { 'c-footer-row--rightAlignedOnDesktopView': !showCountrySelector }]">
+                     { 'c-footer-row--rightAlignedAboveWide': !showCountrySelector }]">
             <country-selector
                 v-if="showCountrySelector"
                 data-test-id="country-selector"
@@ -56,7 +56,7 @@
 
             <legal-field
                 v-if="metaLegalFieldEnabled"
-                :class="[{ 'c-footer-row-item--fullWidthOnDesktopView': !showCountrySelector }]"
+                :class="[{ 'c-footer-row-item--fullWidthAboveWide': !showCountrySelector }]"
                 :info="copy.metaLegalField" />
 
             <icon-list
@@ -215,7 +215,7 @@ $footer-heading-font-size: 'heading-s';
     padding-bottom: 0;
 }
 
-.c-footer-row--rightAlignedOnDesktopView {
+.c-footer-row--rightAlignedAboveWide {
     justify-content: flex-end;
 
     @include media('<wide') {
@@ -223,7 +223,7 @@ $footer-heading-font-size: 'heading-s';
     }
 }
 
-.c-footer-row-item--fullWidthOnDesktopView {
+.c-footer-row-item--fullWidthAboveWide {
     @include media('>=wide') {
         flex: 1;
     }