+++ packages/components/organisms/f-checkout/src/store/checkout.module.js
@@ -81,7 +81,7 @@ export default {
          * @param {Object} payload - Parameter with the different configurations for the request.
          */
         // eslint-disable-next-line no-unused-vars
-        patchCheckout: async ({ commit, state }, {
+        updateCheckout: async ({ commit, state }, {
             url, data, timeout
         }) => {
             // TODO: deal with exceptions and handle this action properly (when the functionality is ready)