+++ packages/components/organisms/f-header/CHANGELOG.md
@@ -4,6 +4,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+To be added to the latest release
+------------------------------
+*Fabruary 9, 2021*
+
+### Added
+- All tenants to a tenant prop for storybook f-header story
+
+
 v4.6.0
 ------------------------------
 *January 27, 2021*